requirements2.txt
future==0.18.2
matplotlib==3.5.1
mediapipe==0.8.11
numpy==1.22.1
omegaconf==2.1.1
opencv_python_headless==4.5.5.62
pandas==1.4.0
Pillow==9.1.1
preprocessing==0.1.13
protobuf==3.20.2
seaborn==0.11.2
setuptools==59.5.0
tensorboard==2.8.0
torch==1.11.0
torchmetrics==0.8.0
torchvision==0.12.0
tqdm==4.64.1
keyboard
pyautogui